vscode support iconPath in QuickPickItem
fixes #12883
VS Code 1.81 finalized the support for the iconPath in QuickPickItem. This is a partial support: full API support and the display of the ThemeIcon, but the display of the URI based icons is not yet available. This last part requires an update in monaco editor core

Contributed on behalf of STMicroelectronics