Fix electron extension host
Before `yarn electron start` was invoked.
This does not work, as the cli tools will not be available when bundled.
Adjust the `getStartCommand` in `AbstractHostedInstanceManager` to work
for both the browser and electron case.
Additionally, add `--electronUserData` to the electron case.
This will avoid odd behavior between the two instances running.
Use a timestamp folder in the tmp folder of the os in order to keep things
clean on the machine and between different debugging sessions.