Add multi conversation support
Added multi conversation support to the chat window.
A new add button can be used to create a new chat session (this can also be done programatically).
With a history button the user can switch to another chat session.
This is done via a quick Picker that shows all other sessions that you can switch to.
Changed the PR finalization feature so that it opens up a new chat session.