chat ui input: Use monaco and add autocomplete for agents and variables
Improve the AI chat ui input field with autocompletion for chat agents
and ai variables:

Typing `@` allows to autocomplete chat agents.
Typing `#` allows to autocomplete ai variables.

To facilitate this, the input field was changed from a textarea to an
inline monaco editor. A custom language `ai-chat-view-language` is
introduced to provide the completions via completion item providers
that query the ChatAgentService resp. AIVariableService

Co-authored-by: Alexandra Muntean <amuntean@eclipsesource.com>
Co-authored-by: Martin Fleck <mfleck@eclipsesource.com>