Chat view ought to be a proper singleton
- restore the singleton scope in the Inversify binding
- do not dispose nested views on disposal of the Chat view
- handle other dependencies on attach/detach of the view

Note that this does result in odd
behaviour when opening the view again
in the main window, yanking it out of\the separate window, as then closing
the separate window closes the view
that is now in the main window. This
boils down to a known issue in the
general support for secondary windows.