Parse functions in prompt templates and add to LLM calls
Extends the prompt service to parse function descriptions in format
`~{<id>}` from prompt templates and return them as part of the
resolved prompt.

To support returning tools besides the plain text for resolved prompts, the message signature of `PromptService.getPrompt`
now returns an object of new type ResolvedPromptTemplate.

Similarly, the `getSystemMessage` method of the`AbstractChatAgent` now also return an object.
The `AbstractChatAgent` is adapted to hand over system and
request function descriptions to the LLM.